+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.1.0 (February 2, 2024)
+## 0.1.0 
 
 - Created the nfs-data-dictionary data dictionary repository using Spout v1.0.0
+- Add initial set of demographic variables
+- Add initial set of self-reported outcomes variables
+- Add harmonized demographic variables
+- The CSV dataset is located here:
+    - \\rfawin.partners.org\bwh-sleepepi-nsrr-staging\20240130-chee-need-for-sleep\nsrr-prep\_releases\0.1.0\